+++ src/mandelbrot.c
@@ -1,15 +1,15 @@
 
 #include "../fractol.h"
 
-int		ft_man_iter(t_env *env, int x, int y)
+int		ft_man_iter(t_env *env, long long x, long long y)
 {
 	t_ftl	ftl;
 	int		i;
 
 	ftl.c_r = (long double)x / env->zoom;
 	ftl.c_i = (long double)y / env->zoom;
-	ftl.z_r = 0;
-	ftl.z_i = 0;
+	ftl.z_r = 0.0;
+	ftl.z_i = 0.0;
 	i = -1;
 	while (++i < env->i_max)
 	{
